refactor: logic for generator separated in a new component called Generator
Francisco Zambrano
